\name{asymscalexample}
\alias{asymscalexample}
\docType{data}
\title{
Asymscal Example Data
}
\description{
This is an artificial dataset. The data are distances from a two-dimensional model, and because of this construction the asymscal model fit this data exactly. In addition, two rows of this matrix have weights different from (1,1). The fifth subject has weights (1.35,.25), and the 15th subject has weights (1.65,.425).
}
\usage{data("asymscalexample")}
\format{
  A matrix with 15 rows and 15 columns.
}
\keyword{datasets}
