\name{Englishtowns}
\alias{Englishtowns}
\docType{data}
\title{
Distance Matrix of Eight English Towns
}
\description{
 A data matrix with 8 rows and 8 columns. The data are distances between eight English towns, this data matrix is made asymmetric by addding linear skew-symmetric matrix. In this dataset, asymmetry is imposed by perturbing the data.
}
\usage{data("Englishtowns")}
\references{
Constantine, A.G. & Gower, J.C. (1978). Graphical Representation of Asymmetric Matrices. Appl. Statist, 27, 297-304.
}
\examples{
data(Englishtowns)
}
\keyword{Datasets}
