\name{print.mdsunique}
\alias{print.mdsunique}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print Method for the MDS Model with unique dimensions
}
\description{
Print method for objects of class mdsunique.
}
\usage{

\method{print}{mdsunique}(x, ...)

}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
Object of class mdsunique
}
  \item{\dots}{
Further arguments
}
}
