\name{productsafety}
\alias{productsafety}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Intercountry Notification of Unsafe Products
}
\description{
The Rapid Alert System for dangerous non-food products (RAPEX) notifies EU member states about risks of products to the health and safety of consumers. Risks for the consumer include choking, strangulation and fire, to name just a few.
Examples of products in this database are powerbanks, clothing, toys, lighters, among others.
Dozens of products in the EU are withdrawn from the market every month because they pose a risk to users health and safety. Market surveillance authorities in EU member states are expected to inform other countries about dangerous products, so that they are removed from the market in other countries. These data are maintained in an exchange system known as RAPEX. Countries can register unsafe products in the RAPEX database, this process is called notification. Other countries may then act on a notification made by one of the other countries.
This table is derived from the RAPEX database. The entries in the table give the number of products removed from the row country, that is acted upon by the column country.
}
\references{
\href{https://english.rekenkamer.nl/publications/reports/2017/01/19/products-sold-on-the-european-market-unraveling-the-system-of-ce-marking}{https://english.rekenkamer.nl/publications/reports/2017/01/19/products-sold-on-the-european-market-unraveling-the-system-of-ce-marking}
}
\keyword{Datasets}% use one of  RShowDoc("KEYWORDS")
