\name{summary.mdsunique}
\alias{summary.mdsunique}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summary method for the mdsunique model
}
\description{
This function extracts the configuration matrix from this multidimensional scaling object.
}
\usage{
\method{summary}{mdsunique}(object, ...)

}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
An object of class \code{mdsunique}
}
  \item{\dots}{
Further arguments
}
}
\keyword{Methods}
