\name{asymmetry-package}
\alias{asymmetry-package}
\alias{asymmetry}
\docType{package}
\title{
\packageTitle{asymmetry}
}
\description{
\packageDescription{asymmetry}
}

\author{
\packageAuthor{asymmetry}

Maintainer: \packageMaintainer{asymmetry}
}
\references{
Zielman, B., and Heiser, W. J. (1993), The analysis of asymmetry by a slide-vector, Psychometrika, 58, 101-114.
}


