\name{Englishtowns}
\alias{Englishtowns}
\docType{data}
\title{
Distance matrix of eight english towns
}
\description{
 A data matrix with 8 rows and 8 columns. The data are a  distance matrix between eight English towns. The data are not symmetric because a linear skew-symmetric matrix is added to the true distances
}
\usage{data("Englishtowns")}
\references{
Constantine, A.G. & Gower, J.C. (1978). Graphical Representation of Asymmetric Matrices. Appl. Statist, 27, 297-304.
}
\examples{
data(Englishtowns)
}
\keyword{datasets}
