\name{print.slidevector}
\alias{print.slidevector}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print Method for the Slide-vector Model
}
\description{
Print method for objects of class slidevector.
}
\usage{

\method{print}{slidevector}(x, ...)

}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
Object of class slidevector
}
  \item{\dots}{
Further arguments
}
}
\keyword{Methods}
